7042 rework created new class for both the response and updated method name AB#7042 AB#9639
